+++ app/src/main/java/com/fibelatti/photowidget/widget/PhotoWidgetAlarmManager.kt
@@ -2,38 +2,69 @@ package com.fibelatti.photowidget.widget
 
 import android.app.AlarmManager
 import android.content.Context
-import android.os.SystemClock
+import androidx.core.app.AlarmManagerCompat
 import androidx.core.content.getSystemService
 import dagger.hilt.android.qualifiers.ApplicationContext
-import timber.log.Timber
-import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 import javax.inject.Singleton
+import timber.log.Timber
 
 @Singleton
 class PhotoWidgetAlarmManager @Inject constructor(
     @ApplicationContext private val context: Context,
+    private val photoWidgetStorage: PhotoWidgetStorage,
 ) {
 
     private val alarmManager: AlarmManager by lazy { requireNotNull(context.getSystemService()) }
 
-    fun setup(appWidgetId: Int, repeatInterval: Long, timeUnit: TimeUnit) {
+    fun setup(appWidgetId: Int) {
         Timber.d("Setting alarm for widget (appWidgetId=$appWidgetId)")
 
-        val interval = timeUnit.toMillis(repeatInterval)
+        val widgetInterval = photoWidgetStorage.getWidgetInterval(appWidgetId = appWidgetId)
+        val intervalMillis = widgetInterval.timeUnit.toMillis(widgetInterval.repeatInterval)
 
+        if (AlarmManagerCompat.canScheduleExactAlarms(alarmManager)) {
+            try {
+                Timber.d("Permission was granted, scheduling exact alarm")
+
+                alarmManager.setExact(
+                    /* type = */ AlarmManager.RTC,
+                    /* triggerAtMillis = */ System.currentTimeMillis() + intervalMillis,
+                    /* operation = */ PhotoWidgetProvider.flipPhotoPendingIntent(
+                        context = context,
+                        appWidgetId = appWidgetId,
+                        rescheduleAlarm = true,
+                    ),
+                )
+            } catch (_: SecurityException) {
+                Timber.d("SecurityException: fallback to inexact alarm")
+
+                setRepeatingAlarm(intervalMillis = intervalMillis, appWidgetId = appWidgetId)
+            }
+        } else {
+            Timber.d("Permission was not granted, scheduling inexact alarm")
+
+            setRepeatingAlarm(intervalMillis = intervalMillis, appWidgetId = appWidgetId)
+        }
+    }
+
+    private fun setRepeatingAlarm(intervalMillis: Long, appWidgetId: Int) {
         alarmManager.setRepeating(
-            /* type = */ AlarmManager.ELAPSED_REALTIME,
-            /* triggerAtMillis = */ SystemClock.elapsedRealtime() + interval,
-            /* intervalMillis = */ interval,
-            /* operation = */ PhotoWidgetProvider.flipPhotoPendingIntent(context, appWidgetId),
+            /* type = */ AlarmManager.RTC,
+            /* triggerAtMillis = */ System.currentTimeMillis() + intervalMillis,
+            /* intervalMillis = */ intervalMillis,
+            /* operation = */ PhotoWidgetProvider.flipPhotoPendingIntent(context = context, appWidgetId = appWidgetId),
         )
     }
 
     fun cancel(appWidgetId: Int) {
         Timber.d("Cancelling alarm for widget (appWidgetId=$appWidgetId)")
-        alarmManager.cancel(
-            /* operation = */ PhotoWidgetProvider.flipPhotoPendingIntent(context, appWidgetId),
+
+        val pendingIntent = PhotoWidgetProvider.flipPhotoPendingIntent(
+            context = context,
+            appWidgetId = appWidgetId,
         )
+
+        alarmManager.cancel(pendingIntent)
     }
 }