+++ app/src/main/java/com/fibelatti/photowidget/widget/PhotoWidgetProvider.kt
@@ -17,11 +17,12 @@ import com.fibelatti.photowidget.di.entryPoint
 import com.fibelatti.photowidget.model.PhotoWidget
 import com.fibelatti.photowidget.model.PhotoWidgetAspectRatio
 import com.fibelatti.photowidget.model.PhotoWidgetTapAction
+import com.fibelatti.photowidget.platform.intentExtras
 import com.fibelatti.photowidget.platform.withPolygonalShape
 import com.fibelatti.photowidget.platform.withRoundedCorners
+import kotlin.math.sqrt
 import kotlinx.coroutines.launch
 import timber.log.Timber
-import kotlin.math.sqrt
 
 class PhotoWidgetProvider : AppWidgetProvider() {
 
@@ -51,11 +52,13 @@ class PhotoWidgetProvider : AppWidgetProvider() {
         if (TAP_TO_FLIP_ACTION == intent.action) {
             runCatching {
                 val entryPoint = entryPoint<PhotoWidgetEntryPoint>(context)
-                val scope = entryPoint.coroutineScope()
-                val flipPhotoUseCase = entryPoint.flipPhotoUseCase()
 
-                scope.launch {
-                    flipPhotoUseCase(appWidgetId = intent.appWidgetId)
+                entryPoint.coroutineScope().launch {
+                    entryPoint.flipPhotoUseCase().invoke(appWidgetId = intent.appWidgetId)
+                }
+
+                if (intent.rescheduleAlarm) {
+                    entryPoint.photoWidgetAlarmManager().setup(appWidgetId = intent.appWidgetId)
                 }
             }
         }
@@ -67,6 +70,8 @@ class PhotoWidgetProvider : AppWidgetProvider() {
 
         // RemoteViews have a maximum allowed memory for bitmaps
         private const val MAX_WIDGET_BITMAP_MEMORY = 6_912_000
+
+        var Intent.rescheduleAlarm: Boolean by intentExtras(default = false)
 
         fun ids(context: Context): List<Int> = AppWidgetManager.getInstance(context)
             .getAppWidgetIds(ComponentName(context, PhotoWidgetProvider::class.java))
@@ -182,15 +187,17 @@ class PhotoWidgetProvider : AppWidgetProvider() {
         fun flipPhotoPendingIntent(
             context: Context,
             appWidgetId: Int,
+            rescheduleAlarm: Boolean = false,
         ): PendingIntent {
-            val clickIntent = Intent(context, PhotoWidgetProvider::class.java).apply {
+            val intent = Intent(context, PhotoWidgetProvider::class.java).apply {
                 this.appWidgetId = appWidgetId
                 this.action = TAP_TO_FLIP_ACTION
+                this.rescheduleAlarm = rescheduleAlarm
             }
             return PendingIntent.getBroadcast(
                 /* context = */ context,
                 /* requestCode = */ appWidgetId,
-                /* intent = */ clickIntent,
+                /* intent = */ intent,
                 /* flags = */ PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT,
             )
         }